Move comparison function out of utils/tests which is for tests of utility functions
